Refactor common methods from gateway and wasm tests into util
