+++ test/kube2e/util.go
@@ -2,14 +2,23 @@ package kube2e
 
 import (
 	"context"
+	clienthelpers "github.com/solo-io/gloo/projects/gloo/cli/pkg/helpers"
 	"io/ioutil"
+	"net/http"
 	"os"
+	"os/exec"
+	"regexp"
+	"strconv"
+	"time"
 
+	"github.com/gogo/protobuf/types"
 	. "github.com/onsi/gomega"
 	errors "github.com/rotisserie/eris"
 	"github.com/solo-io/gloo/projects/gloo/cli/pkg/cmd/check"
 	"github.com/solo-io/gloo/projects/gloo/cli/pkg/cmd/options"
+	v1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
 	"github.com/solo-io/go-utils/testutils/helper"
+	"github.com/solo-io/solo-kit/pkg/api/v1/clients"
 	"github.com/solo-io/solo-kit/pkg/api/v1/resources/core"
 
 	"github.com/solo-io/go-utils/kubeutils"
@@ -74,4 +83,86 @@ gloo:
 	Expect(err).NotTo(HaveOccurred())
 
 	return values.Name(), func() { _ = os.Remove(values.Name()) }
+}
+
+func EventuallyReachesConsistentState(installNamespace string) {
+	metricsPort := strconv.Itoa(9091)
+	portFwd := exec.Command("kubectl", "port-forward", "-n", installNamespace,
+		"deployment/gloo", metricsPort)
+	portFwd.Stdout = os.Stderr
+	portFwd.Stderr = os.Stderr
+	err := portFwd.Start()
+	Expect(err).ToNot(HaveOccurred())
+
+	defer func() {
+		if portFwd.Process != nil {
+			portFwd.Process.Kill()
+		}
+	}()
+
+	// make sure we eventually reach an eventually consistent state
+	lastSnapOut := getSnapOut(metricsPort)
+
+	eventuallyConsistentPollingInterval := 7 * time.Second // >= 5s for metrics reporting, which happens every 5s
+	time.Sleep(eventuallyConsistentPollingInterval)
+
+	Eventually(func() bool {
+		currentSnapOut := getSnapOut(metricsPort)
+		consistent := lastSnapOut == currentSnapOut
+		lastSnapOut = currentSnapOut
+		return consistent
+	}, "30s", eventuallyConsistentPollingInterval).Should(Equal(true))
+
+	Consistently(func() string {
+		currentSnapOut := getSnapOut(metricsPort)
+		return currentSnapOut
+	}, "30s", eventuallyConsistentPollingInterval).Should(Equal(lastSnapOut))
+}
+
+// needs a port-forward of the metrics port before a call to this will work
+func getSnapOut(metricsPort string) string {
+	var bodyResp string
+	Eventually(func() string {
+		res, err := http.Post("http://localhost:"+metricsPort+"/metrics", "", nil)
+		if err != nil || res.StatusCode != 200 {
+			return ""
+		}
+		defer res.Body.Close()
+		body, err := ioutil.ReadAll(res.Body)
+		Expect(err).ToNot(HaveOccurred())
+		bodyResp = string(body)
+		return bodyResp
+	}, "3s", "0.5s").ShouldNot(BeEmpty())
+
+	Expect(bodyResp).To(ContainSubstring("api_gloo_solo_io_emitter_snap_out"))
+	findSnapOut := regexp.MustCompile("api_gloo_solo_io_emitter_snap_out ([\\d]+)")
+	matches := findSnapOut.FindAllStringSubmatch(bodyResp, -1)
+	Expect(matches).To(HaveLen(1))
+	snapOut := matches[0][1]
+	return snapOut
+}
+
+// enable/disable strict validation
+func UpdateAlwaysAcceptSetting(alwaysAccept bool, installNamespace string) {
+	UpdateSettings(func(settings *v1.Settings) {
+		Expect(settings.Gateway).NotTo(BeNil())
+		Expect(settings.Gateway.Validation).NotTo(BeNil())
+		settings.Gateway.Validation.AlwaysAccept = &types.BoolValue{Value: alwaysAccept}
+	}, installNamespace)
+}
+
+func UpdateSettings(f func(settings *v1.Settings), installNamespace string) {
+	settingsClient := clienthelpers.MustSettingsClient()
+	settings, err := settingsClient.Read(installNamespace, "default", clients.ReadOpts{})
+	Expect(err).NotTo(HaveOccurred())
+
+	f(settings)
+
+	_, err = settingsClient.Write(settings, clients.WriteOpts{OverwriteExisting: true})
+	Expect(err).NotTo(HaveOccurred())
+
+	// when validation config changes, the validation server restarts -- give time for it to come up again.
+	// without the wait, the validation webhook may temporarily fallback to it's failurePolicy, which is not
+	// what we want to test.
+	time.Sleep(3 * time.Second)
 }
