+++ test/kube2e/wasm/wasm_test.go
@@ -1,4 +1,4 @@
-package gateway_test
+package wasm_test
 
 import (
 	"context"
